+++ src/events/files.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { RCECancelEvent } from './utils';
-import { CONFIG } from '../config';
+import { CONNECTION } from '../config';
 import { getWorkspaceUri } from '@/utils';
 
 /**
@@ -11,8 +11,8 @@ import { getWorkspaceUri } from '@/utils';
  */
 export function targetedFileCreatedEvent(file: string) {
     return new RCECancelEvent<vscode.FileCreateEvent>({
-        reason: `the file ${file} was created by Vedal.`,
-        logReason: (_data) => `${CONFIG.currentlyAsNeuroAPI} created the file ${file}.`,
+        reason: `the file ${file} was created by ${CONNECTION.userName}.`,
+        logReason: () => `the file "${file}" was created.`,
         events: [
             [vscode.workspace.onDidCreateFiles, (data) => {
                 const workspaceUri = getWorkspaceUri();
@@ -31,8 +31,8 @@ export function targetedFileCreatedEvent(file: string) {
  */
 export function targetedFileDeletedEvent(file: string) {
     return new RCECancelEvent<vscode.FileDeleteEvent>({
-        reason: `the file ${file} was deleted by Vedal.`,
-        logReason: (_data) => `${CONFIG.currentlyAsNeuroAPI} deleted the file ${file}.`,
+        reason: `the file ${file} was deleted by ${CONNECTION.userName}.`,
+        logReason: () => `the file "${file}" was created.`,
         events: [
             [vscode.workspace.onDidDeleteFiles, (data) => {
                 const workspaceUri = getWorkspaceUri();